refactor: move updateFinancialRecords to its own file
